pyproject.toml
[project]
name = "flytbase"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.12.0",
    "bitsandbytes>=0.49.0",
    "boto3>=1.42.21",
    "chromadb>=1.4.0",
    "fastapi>=0.128.0",
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "groq>=1.0.0",
    "hf-xet>=1.2.0",
    "opencv-python>=4.12.0.88",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.52.2",
    "torch>=2.9.1",
    "transformers>=4.57.3",
    "uvicorn>=0.40.0",
]
